requirements.txt
flask==3.1.1
pymongo==4.12.1
python-dotenv==1.1.0
pyjwt==2.10.1
bcrypt==4.3.0
flask-cors==5.0.1
twilio==9.5.1
pytest==8.3.5
pytest-cov==6.1.1
